everything is restructured
to make it a lot more organized
